+++ main.py
@@ -1,0 +1,3 @@
+import scripts.taptap as taptap
+
+taptap.TapTapClient().check_latest()