+++ CHANGELOG.md
@@ -6,7 +6,8 @@
 ### [未发布] -- 2026-04-09
 #### 新增:
 - [taptap.py](./scripts/taptap.py) -- 来源 [![Phigros_Resource](https://img.shields.io/badge/GitHub-Phi--Resource-red?logo=github)](https://github.com/7aGiven/Phigros_Resource/blob/master/taptap.py)
-- [CHANGELOG_EN.md](./CHANGELOG_EN.md)
+- [CHANGELOG_EN.md](./CHANGELOG_EN.md) 英文版
+- [downloader.py](./scripts/downloader.py) 勉强可用
 
 #### 修改：
 - 用 `class` 重写 [taptap.py](./scripts/taptap.py) 并添加 **没什么用的** 版本对比功能 -- 来源 [![Phigros_Resource](https://img.shields.io/badge/GitHub-Phi--Resource-red?logo=github)](https://github.com/7aGiven/Phigros_Resource/blob/master/taptap.py)